calculate percentages when there are one or two numbers supplied
if there is one number and it has a % sign, pressing = will calculate that number / 100. If there are two numbers and only one has a percent sign, the one with a percent sign will be converted to that percentage of the other number. Then the calculation will be performed with the new calculated percentage and the non-percent number. If both numbers have % signs, they will be calculated as regular numbers and the result will have a % sign (the % sign is not currently displayed).